build: remove `DEPLOYMENT_ENABLE_LIBDISPATCH`
This flag is no longer needed as a C flag since the prefix header in
CoreFoundation no longer relies on this macro.  This allows us to remove
it from the other builds as CoreFoundation inclusion would not cause any
issues now.  The macro can be fully limited to just the Foundation
build.