CoreFoundation: avoid `CFString` construction in `CFURLSessionInterface`
This causes a reference to `CFStringCreateWithBytes` and
`kCFDefaultSystemAllocator` in `CFURLSessionInterface` which requires
linking against CoreFoundation.  Since the swift deployment uses
CoreFoundation statically, this would require a second instance of
CoreFoundation OR that private CoreFoundation interfaces are re-exposed
through Foundation which is explicitly undesired.  Perform the String
construction in Swift and instead just have a trivial wrapper for
`curl_easy_strerror`.