Base: stub KnownLocations for android
Add a path through KnownLocations for android.  This currently cannot be
computed as we do not have the activity context nor do we know if the
activity was spawned from the zygote or was a NDK process so we do not
have a reference to Davlik to JNI and query the required information.
This is sufficient to get the SDK building for android though.