Improve cookie parsing algorithm.
While the previous parser dealt with commas, some cases were not
correctly handled, like when a comma was in the last attribute before a
second cookie.

Also, improve the cookie parsing algorithm to match better the RFC. The
parser is slightly more strict, and will not allow some malformed
cookies to happen anymore.

Also trim the names of the cookies to allow spaces before the equal.

Add some more tests to improve the coverage and ensure the added pieces
are working as supposed.