[Windows] Fix TestURL.swift crash
On Windows, `|` is a valid character in a URL (as a replacement to `:`,
e.g. `/C|/Users/gwenm`. This causes a test testing
`http://test.com/unescaped|pipe` which is expected  to not parse to
parse successfully on Windows. The test then tries to parse the `<null
url>` in the expected result and crashes.

This changes the test to check to make sure it can parse the expected
result instead of force casting so that the test fails instead of
crashing.

This allows TestURL to run to completion in the mean time while working
on URL for Windows.