[android] Fix symlink comparison in Android.
The size reported by stat.st_size might not be the size of the contents
of the symlink, since it might include metadata. In Android at least the
size of a symlink seems to start at 18 bytes, even if the contents are
smaller.

The commit disables the st_size comparison against the contents
comparison for Android, and rewrites a little the function to use Swift
constructs like defer and guards.