Link correctly to zlib when using cURL statically.
Since we are not using target_link_libraries, it seems that CMake is not
automatically linking to the cURL dependencies when cURL is compiled
statically. This will mean that libFoundationNetworking might have
undefined symbols and cannot be used.

Since the tests check for a gzipped request/response and no modern
client should not support gzipped responses, the find_package calls
require the zlib component for cURL. Additionally, in the case of being
a static library, libz is linked in the cURL dependencies.

The cURL libraries are also removed as interface dependencies of
Foundation, since the only library that depends on cURL is
FoundationNetworking.

There's also a little bit of movement around LIBXML2_LIBRARIES since
they are suppose to come from CoreFoundation_LIBRARIES, so no need to
repeat them.