Fix CFInfo on 64-bit big endian platforms.
Prior to this change 64-bit platforms treated _cfinfoa as both a
64-bit and 32-bit variable. When accessing _cfinfoa as a 32-bit
variable the code makes the assumption that it is accessing the
lowest (or rightmost) bits of _cfinfoa. Unfortunately on big-endian
platforms it is actually accessing the high (or leftmost) bits.

Fix this issue by unconditionally treating _cfinfoa as a 64-bit
value when targeting a 64-bit platform.