Place read-only data symbols in the .rodata section in ELF objects.
This data was previously put into the default section, .text, which
is generally used for executable data. Currently this is only fixed
when using ELF but we should probably fix this for other object file
formats too.

This fixes a gold warning on IBM Z: "S/390 code fill of odd length
requested". This warning was triggered because executable
instructions on IBM Z must be 2-byte aligned and the data in some of
these symbols is 1-byte aligned.