build: workaround hook for zlib dependency
Until CMake 3.15, dependency tracking is not properly possible due to custom
commands for linking.  Provide an escape hatch in the form is a variable that
can be setup to force the linking of zlib.  We do not search for zlib, simply
allow the user specify a magic variable to permit them to inject the linking.