build: restructure CoreFoundation build rules
Reduce some of the unnecessary duplication in the CoreFoundation build.
This also reorganises the rules some to group the operations together.
There is more than can be done here to reduce duplication and simplify
the rules, but some of that work requires coordinating changes with
libdispatch and can be done subsequently.