[Windows] Match error behavior of POSIX
On POSIX Foundation, removing  a file is considered writing to it and so
when the file to be removed isn't found, it throws fileNoSuchFile. Since
Windows instead checks for the file attributes first which is considered
a read, it throws there with fileReadNoSuchFile. In removeItem to match
that behavior, we catch the fileReadNoSuchFile and turn it into
fileNoSuchFile.