SR-11766: Fix coefficients for UnitVolume
This fixes Measurement values expressed in cubic centimeters and cubic millimeters.

I found the errors in the coefficients because I tested one of my libraries on Linux for the first time and a test was failing (see https://github.com/ole/Ampere/issues/25).

I considered adding unit tests that test some conversions to Foundation, but I'm not sure how useful they would be. A test for converting a Measurement from one unit to another would essentially repeat the code that's being tested; it seems it would be almost equally likely for a bug to be in the test as it would be in the code being tested.