[CF] DT_WHT is unavailable on OpenBSD.
Here we duplicate this extended conditional to not reference DT_WHT,
which is not available on OpenBSD. It is probably harmless to actually
remove DT_WHT entirely here unconditionally, but would require grokking
more context for a simple buildfix.

It also seems more readable to duplicate the entire line in the ifdef
despite being able to squish the ifdef in the clause.