[CF] flsl is required for OpenBSD.
flsl is available on FreeBSD (and thus macOS), but not OpenBSD. This
commit makes the implementation of flsl currently exposed to Linux and
Windows also available to OpenBSD. Thus, we can move this implementation
up to the existing block of shared infrastructure.

Additionally, this commit separates it from the block below which
declares asprintf. The asprintf declaration is still required for Linux
and Windows, but not OpenBSD.

While we are here, move an existing comma loop body to explicit braces.