[CF] Do not call malloc_size unless debugging.
CFBasicHashGetSize has an argument, total, to enable whether to
presumably calculate the size its dependent data structures.
CFBasicHashGetSize does not appear outside CFBasicHash.c, and this total
argument is true only when ENABLE_MEMORY_COUNTERS is true or
transitively when ENABLE_DTRACE_PROBES is true. These preprocessor
symbols appear to be to control debugging and other informational
purposes.

malloc introspection, as afforded by malloc_size, malloc_usable_size, or
msize, is nonportable (and thus not guaranteed to be present on all
platforms), not guaranteed to be accurate (though certainly an upper
bound), and is generally recommended for debugging.

For those reasons, only expose the call to malloc_size when either of
ENABLE_MEMORY_COUNTERS or ENABLE_DTRACE_PROBES is true. This avoids an
undefined symbol issue when malloc introspection is unavailable, but
does mean these counters and other information is not available on
platforms when malloc_size is not present.