[CF] Remove __CFStorageEstimateTotalAllocatedSize.
This function is tagged with __attribute__((unused)) and certainly
appears to be unused. It may be a function intended for use in a
debugger context. As it is a static function, the reference to this
function and malloc_size will not necessarily be present when built,
_but_ this does cause a compiler warning when building on systems where
malloc_size or alternatives are not available.

malloc introspection, as afforded by malloc_size, malloc_usable_size, or
msize, is nonportable (and thus not guaranteed to be present on all
platforms), not guaranteed to be accurate (though certainly an upper
bound), and is generally recommended for debugging.

For those reasons, remove this function.