[CF] Move pthread types to ForFoundationOnly.h.
CFPlatform.c has dependencies on pthreads, such as in _CFIsMainThread
and the type of __CFTSDIndexKey, so the pthread header needs to be
included somewhere and the relevant typedefs for _CFThreadRef and others
need to be present.

One place this could be done is in CoreFoundation_Prefix.h, however, we
find the necessary definitions for Windows and pthread environments
present in ForSwiftFoundationOnly.h. Unfortunately, as these
declarations are necessary outside of the Swift context, this leads us
to ForFoundationOnly.h, since that is included (in CFInternal.h) for
_both_ Swift and non-Swift builds of CoreFoundation.

To remove duplication, the unistd.h and pthread.h checks are removed
from ForSwiftFoundationOnly.h, because these are also found in
CFInternal.h.