FileManager: correct symbolic link handling on Windows
When the symbolic link points across a drive, querying the symbolic link
will provide a NT style path.  The NT style path is not valid for a
reparse point destination.  Use a hand-rolled canonicalization (removal
of the `\??\` prefix) for the path to get a path that we can use in the
symbolic link.  `NTPathToDosPath` is a possible approach to a more
robust solution but relies on undocumented kernel APIs.