Foundation: fix missing string terminator
The computation of the base64encoding would trim the trailing nul
terminator on the rendered string.  This would result in invalid
string manipulation leading to heap corruptions on Windows.

Thanks to Simon Evans for spotting the issue and proposing the fix!
(And a thank you for spotting a leak that I was accidentally
introducing.)