Revert "Foundation: fix missing string terminator"
This reverts commit f3a6f50105ae4e922cc4146a25eb1a9313992a98.  It also
fixes the underlying issue of the heap corruption.

There is no guarantee that the underlying storage for an allocation is
done by `malloc`.  It is possible that the size may be small enough to
be pushed down to a stack allocation, which cannot be released by
`free`.  Buffers record how they are allocated and adjust the release
accordingly.  The safe way to release the memory is to use `deallocate`.

Although the initial approach fixed the heap corruption, it did so at
the single site, but this is a problem with a public interface.  Use the
`deallocate` method on the buffer when releasing the memory to ensure
that the correct mechanism is used for releasing the resources.

This is not specific to Windows but to any platform where a stack
allocated buffer or an aligned allocation may not be passed to `free`.

Thanks to Lily Vulcano for pressing me to investigate this further and
root cause the failure in the public interface!