TimeZone: delegate to ICU on Windows
Unfortunately, Windows does not have full time zone information
available.  For many cases we can make do with the system information,
augmented with alternate names (the Olson-to-Windows conversion
database).  This papers over the inability to deserialize the full
information in the registry by falling back to ICU to get the
information for the DST time conversion.