[CF] Unify return types in CFPropertyList again.
In #2598, we brought the prototypes and declarations into alignment for
_CFPropertyListCreateFiltered and _CFPropertyListCreateSingleValue,
namely, using a `Boolean` for prototype and declaration, since they were
not matched.

It appears with the Catalina merge they were brought into alignment the
other way around, namely using `bool` instead. Since the Catalina merge
chose `bool`, let's just go with that.