Foundation: extract `cfSource` from `RunLoop._Source`
Making `cfSource` a member within the structure prevents allowing
overriding of the `open` class as the type it returns cannot be
deserialised.  Because the member is not a stored property, extract it
to an extension which removes it from the vtable.  This allows the use
of the private type in an external context.