Remove an invalid numeric JSON test
The value `2.7976931348623158e+308` is a perfectly valid JSON number
and should be accepted.  This test is only passing today because the Swift
standard library's `Double(_ String:)` initializer has a bug that
rejected numbers larger than the maximum Double value.

But we're getting ready to fix that bug in the Swift standard library,
(apple/swift#34339) which will break this test when it lands.