Store the hostname in _names for Host.current()
Host.current() looks up the hostname for its _info field, but then never
gives it to the user. As a result, .names is always empty for
Host.current().

This patch is a minimal-effort workaround: it simply throws the value of
gethostname() (or its Windows equivalent) into _names. We could probably
do better than that, but for now this is good enough.