Process: handle `HRESULT` exit codes properly
When a process terminates due to an unhandled exception via a `HRESULT`,
the process termination status will not have `0xC000_0000` or
`0xE000_0000` but rather `0x8000_0000` set.  This accounts for that and
properly classifies the exit as an abnormal exit.  This was identified
by the swift-driver test suite not properly detecting an abnormal exit
of the frontend.