build: allow static linking against ICU data
With the removal of the ICU data dependency in swiftCore, it is now
possible to statically link to the ICU, with static linking to the
ICU data as well.  This would allow us to remove the ICU dependency
from leaking into distributions which was previously required due to
the oversized cost of the ICU data.  With the single use now, it is
possible to statically link the data to avoid the distribution and
packaging.  In order to provide a migration path, this needs to be
optionalised, and as such, we only link the data if it is found (and
can be controlled through `-D ICU_DATA_LIBRARY_RELEASE=`