Foundation: make some functions more visible
These ivars and functions are referenced elsewhere in Foundation.  We
were previously relying on the compiler exporting functions more
aggressively which would ensure that these were emitted even when not
meant to be used outside of the current module.  This marks these
interfaces as `internal` to ensure that references from other source
files are emitted with more aggressive internalization by the compiler.