Foundation: correct `Thread.callStackSymbols` on Windows
We would previously print invalid strings due to invalid memory usage.
The string name is not guaranteed to be nul-terminated, and the
`String(cString:)` initializer requires that guarantee.  Explicitly do
the alloc + init approach using `strncpy` to initialize the string in
place. This corrects that and takes the opportunity to simplify some of
the deallocation handling by using a temporary allocation.