Do not chdir before calling posix_spawn
We allow users to set the working directory of the child process when
they use the Process API. Unfortunately, this was implemented by calling
chdir in the parent process, which has nasty global effects on the
process and cannot safely be used concurrently.

glibc has a non-POSIX extension to posix_spawn that can call chdir in
the child process, so we add that call instead where it's available, and
add a regression test.