Base: add support for querying user home directory on Windows
The user profile directory is not possible to access without the user
token for the `GetUserProfileDirectoryW`.  Instead, we can iterate the
users matching a particular filter (e.g. local users).  This allows us
to at least query some user home directory.  Unfortunately, this would
disallow querying of users from a domain controller.  This resolves an
immediate issue of SPM tests depending on the ability to query profile
information for a user.