Remove unnecessary use of renameat2.
On Linux, on some code paths, we call renameat2 with `RENAME_EXCHANGE`
in the implementation of replaceItemAt(_:withItemAt:). This is a
behaviour utterly unique to that code path: Darwin Foundation doesn't do
this, and the non-renameat2 codepaths don't do it either.

As Darwin doesn't do this, we maximise compatibility and minimise
complexity by just deleting the path and all supporting code altogether.

Resolves #4655.