FileManager: repair `homeDirectoryForCurrentUser` on Windows
`NetUserGetInfo` will not provide the home directory for the specified
user most of the time, even if the user is the currently logged in user.
Prefer to use the `CFCopyHomeDirectoryURLForUser` path which will
provide the home directory for the current user.  Fast-path that into
the computed property and add a special case for
`homeDirectory(forUser:)` to accommodate this behavioural difference.