Process: disable Nagle on Windows
Windows does not have a native `socketpair` interface and so we emulate
that locally.  Furthermore, the use of `socketpair` on the other
platforms uses an `AF_UNIX` (Unix Domain Socket) which is not guaranteed
to be available on Windows.  As a result, we workaround that by using an
`AF_INET` address family socket.  This will by default have Nagle
enabled which can potentially cause some delays when processing a
message.  We have observed occasional hangs in the process handling
which may be possibly be attributed to the buffering.  Disable nagle on
the sender side in the hopes of a more reliable connection.